+++ app/src/main/java/com/pzbapps/squiggly/common/domain/utils/Constant.kt
@@ -28,5 +28,7 @@ object Constant {
     const val SORT_ORDER_KEY = "sort_order_key"
     const val SORT_ORDER_VALUE_1 = "created date"
     const val SORT_ORDER_VALUE_2 = "modified date"
+    const val SHOW_RATING_DIALOG_BOX = "show_rating_dialog_box"
+    const val SHOW_RATING_DIALOG_BOX_KEY = "show_rating_dialog_box_key"
 
 }