+++ app/src/main/java/com/pzbapps/squiggly/common/presentation/alertboxes/ratingDialogBox/showRatingDialogBox.kt
@@ -1,0 +1,95 @@
+package com.pzbapps.squiggly.common.presentation.alertboxes.ratingDialogBox
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.shape.CornerSize
+import androidx.compose.material.Button
+import androidx.compose.material.ButtonDefaults
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.OutlinedButton
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.pzbapps.squiggly.common.presentation.FontFamily
+
+@Composable
+fun ShowRatingDialogBox(onDismiss: () -> Unit) {
+    var scope = rememberCoroutineScope()
+    androidx.compose.material3.AlertDialog(
+        onDismissRequest = {
+            onDismiss()
+        },
+        shape = MaterialTheme.shapes.medium.copy(
+            topStart = CornerSize(15.dp),
+            topEnd = CornerSize(15.dp),
+            bottomStart = CornerSize(15.dp),
+            bottomEnd = CornerSize(15.dp),
+        ),
+        containerColor = MaterialTheme.colors.primaryVariant,
+        /*      icon = {
+                     Icon(imageVector = Icons.Filled.Delete, contentDescription = "Delete")
+              }*/
+
+        title = {
+            Text(
+                text = "Rate app 🙏",
+                fontFamily = FontFamily.fontFamilyBold,
+                fontSize = 20.sp,
+                color = MaterialTheme.colors.onPrimary
+            )
+        },
+        text = {
+            Text(
+                text = "Your support means everything to me!  " +
+                        "By giving me a quick 5-star rating, you're " +
+                        "helping me reach more users, " +
+                        "which inspires me to keep improving and " +
+                        "adding amazing features.It only takes 5 seconds." +
+                        " Thank you so much 😊. ",
+
+                fontFamily = FontFamily.fontFamilyRegular,
+                color = MaterialTheme.colors.onPrimary,
+                fontSize = 14.sp
+            )
+        },
+        confirmButton = {
+            Button(
+                onClick = {
+
+                },
+                colors = ButtonDefaults.buttonColors(
+                    backgroundColor = MaterialTheme.colors.onPrimary,
+                    contentColor = MaterialTheme.colors.primary
+                ),
+                shape = MaterialTheme.shapes.medium.copy(
+                    topStart = CornerSize(15.dp),
+                    topEnd = CornerSize(15.dp),
+                    bottomStart = CornerSize(15.dp),
+                    bottomEnd = CornerSize(15.dp),
+                )
+            ) {
+                Text(text = "Rate app", fontFamily = FontFamily.fontFamilyRegular)
+            }
+        },
+        dismissButton = {
+            OutlinedButton(
+                onClick = { onDismiss() },
+                shape = MaterialTheme.shapes.medium.copy(
+                    topStart = CornerSize(15.dp),
+                    topEnd = CornerSize(15.dp),
+                    bottomStart = CornerSize(15.dp),
+                    bottomEnd = CornerSize(15.dp),
+                ),
+                border = BorderStroke(1.dp, MaterialTheme.colors.onPrimary),
+                colors = ButtonDefaults.buttonColors(
+                    backgroundColor = MaterialTheme.colors.primaryVariant,
+                    contentColor = MaterialTheme.colors.onPrimary
+                ),
+            ) {
+                Text(text = "Already rated", fontFamily = FontFamily.fontFamilyRegular)
+            }
+        }
+    )
+}